refactor: split the three remaining large files
Mechanical splits — public API preserved via pub use, no behavior changes.

- src/core/serve_client.rs (1478 → 223) + serve_client/
  ops.rs (442), pipelined.rs (281), pool.rs (334), promote.rs (169),
  test_support.rs (80)

- src/core/remote.rs (1316 → 221) + remote/
  transfer.rs (462), ops.rs (215), attrs.rs (180), resume.rs (148),
  ssh_cmd.rs (135)

- src/main.rs (904 → 182) + src/app/
  commands.rs (408), completions.rs (90), status.rs (87),
  prompts.rs (80), runners.rs (68), updates.rs (28), mod.rs (6)

Visibility widened only where sibling submodules require it; callers use
the same paths (crate::core::serve_client::X, crate::core::remote::X)
via pub use re-exports. test_support submodule gated whole with
#[cfg(any(test, feature = "test-support"))].
