+++ src/app/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod commands;
+pub(crate) mod completions;
+pub(crate) mod prompts;
+pub(crate) mod runners;
+pub(crate) mod status;
+pub(crate) mod updates;
