+++ src/core/remote.rs
@@ -1,13 +1,22 @@
-use crate::core::error::BcmrError;
-use std::path::Path;
-use std::sync::atomic::{AtomicBool, Ordering};
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::process::Command;
+mod attrs;
+mod ops;
+mod resume;
+mod ssh_cmd;
+mod transfer;
 
-static SSH_COMPRESS: AtomicBool = AtomicBool::new(false);
+#[allow(unused_imports)]
+pub use attrs::verify_remote_file;
+#[allow(unused_imports)]
+pub use ops::{
+    complete_remote_path, remote_file_hash, remote_file_size, remote_list_files, remote_stat,
+    remote_total_size, validate_ssh_connection,
+};
+pub use transfer::{
+    download_directory, download_file, ensure_remote_tree, upload_directory, upload_file,
+};
 
 pub fn set_ssh_compression(enabled: bool) {
-    SSH_COMPRESS.store(enabled, Ordering::Relaxed);
+    ssh_cmd::SSH_COMPRESS.store(enabled, std::sync::atomic::Ordering::Relaxed);
 }
 
 #[derive(Debug, Clone)]
@@ -42,10 +51,6 @@ impl std::fmt::Display for RemotePath {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.display())
     }
-}
-
-fn shell_escape(s: &str) -> String {
-    s.replace('\'', "'\\''")
 }
 
 pub fn parse_remote_path(s: &str) -> Option<RemotePath> {
@@ -115,1024 +120,10 @@ pub struct TransferOptions {
 
 pub type RemoteTransferOptions = TransferOptions;
 
-struct ResumeDecision {
-    skip_bytes: u64,
-    use_append_mode: bool,
-    skip_entirely: bool,
-}
-
-async fn check_resume_state(
-    opts: &RemoteTransferOptions,
-    existing_size: Option<u64>,
-    source_size: u64,
-    existing_full_hash: impl AsyncFnOnce() -> Result<String, BcmrError>,
-    source_full_hash: impl AsyncFnOnce() -> Result<String, BcmrError>,
-    source_partial_hash: impl AsyncFnOnce(u64) -> Result<String, BcmrError>,
-) -> Result<ResumeDecision, BcmrError> {
-    if !(opts.resume || opts.append || opts.strict) {
-        return Ok(ResumeDecision {
-            skip_bytes: 0,
-            use_append_mode: false,
-            skip_entirely: false,
-        });
-    }
-
-    let existing_size = match existing_size {
-        Some(s) => s,
-        None => {
-            return Ok(ResumeDecision {
-                skip_bytes: 0,
-                use_append_mode: false,
-                skip_entirely: false,
-            })
-        }
-    };
-
-    if existing_size == source_size {
-        let ex_hash = existing_full_hash().await?;
-        let src_hash = source_full_hash().await?;
-        if ex_hash == src_hash {
-            return Ok(ResumeDecision {
-                skip_bytes: 0,
-                use_append_mode: false,
-                skip_entirely: true,
-            });
-        }
-        return Ok(ResumeDecision {
-            skip_bytes: 0,
-            use_append_mode: false,
-            skip_entirely: false,
-        });
-    } else if existing_size < source_size {
-        let ex_hash = existing_full_hash().await?;
-        let partial = source_partial_hash(existing_size).await?;
-        if ex_hash == partial {
-            return Ok(ResumeDecision {
-                skip_bytes: existing_size,
-                use_append_mode: true,
-                skip_entirely: false,
-            });
-        }
-        return Ok(ResumeDecision {
-            skip_bytes: 0,
-            use_append_mode: false,
-            skip_entirely: false,
-        });
-    }
-
-    Ok(ResumeDecision {
-        skip_bytes: 0,
-        use_append_mode: false,
-        skip_entirely: false,
-    })
-}
-
-fn control_path(target: &str) -> String {
-    let dir = std::env::temp_dir().join("bcmr-ssh");
-    let _ = std::fs::create_dir_all(&dir);
-    dir.join(format!("{}.sock", target.replace(['@', ':', '/'], "_")))
-        .to_string_lossy()
-        .to_string()
-}
-
-fn is_interactive() -> bool {
-    #[cfg(unix)]
-    {
-        unsafe { libc::isatty(libc::STDIN_FILENO) != 0 }
-    }
-    #[cfg(not(unix))]
-    {
-        true
-    }
-}
-
-fn ssh_base_args(target: &str) -> Vec<String> {
-    let cp = control_path(target);
-    let mut args = vec![
-        "-o".into(),
-        format!("ControlPath={}", cp),
-        "-o".into(),
-        "ControlMaster=auto".into(),
-        "-o".into(),
-        "ControlPersist=300".into(),
-        "-o".into(),
-        "ConnectTimeout=10".into(),
-    ];
-    if !is_interactive() {
-        args.extend(["-o".into(), "BatchMode=yes".into()]);
-    }
-    if SSH_COMPRESS.load(Ordering::Relaxed) {
-        args.extend(["-o".into(), "Compression=yes".into()]);
-    }
-    args
-}
-
-fn ssh_base_args_for_worker(target: &str, worker_id: usize) -> Vec<String> {
-    let dir = std::env::temp_dir().join("bcmr-ssh");
-    let _ = std::fs::create_dir_all(&dir);
-    let cp = dir
-        .join(format!(
-            "w{}_{}.sock",
-            worker_id,
-            target.replace(['@', ':', '/'], "_")
-        ))
-        .to_string_lossy()
-        .to_string();
-
-    let mut args = vec![
-        "-o".into(),
-        format!("ControlPath={}", cp),
-        "-o".into(),
-        "ControlMaster=auto".into(),
-        "-o".into(),
-        "ControlPersist=60".into(),
-        "-o".into(),
-        "ConnectTimeout=10".into(),
-    ];
-    if !is_interactive() {
-        args.extend(["-o".into(), "BatchMode=yes".into()]);
-    }
-    if SSH_COMPRESS.load(Ordering::Relaxed) {
-        args.extend(["-o".into(), "Compression=yes".into()]);
-    }
-    args
-}
-
-fn ssh_command_for_worker(target: &str, worker_id: usize) -> Command {
-    let args = ssh_base_args_for_worker(target, worker_id);
-    let mut cmd = Command::new("ssh");
-    for arg in &args {
-        cmd.arg(arg);
-    }
-    cmd.arg(target);
-    cmd
-}
-
-fn make_ssh_cmd(target: &str, worker_id: Option<usize>) -> Command {
-    match worker_id {
-        Some(id) => ssh_command_for_worker(target, id),
-        None => ssh_command(target),
-    }
-}
-
-fn ssh_command(target: &str) -> Command {
-    let args = ssh_base_args(target);
-    let mut cmd = Command::new("ssh");
-    for arg in &args {
-        cmd.arg(arg);
-    }
-    cmd.arg(target);
-    cmd
-}
-
-fn ssh_error_message(stderr: &str, context: &str) -> String {
-    let stderr_lower = stderr.to_lowercase();
-    if stderr_lower.contains("connection refused") {
-        format!(
-            "{}: SSH connection refused (is sshd running on the host?)",
-            context
-        )
-    } else if stderr_lower.contains("no route to host")
-        || stderr_lower.contains("network is unreachable")
-    {
-        format!("{}: host unreachable (check network connectivity)", context)
-    } else if stderr_lower.contains("permission denied") {
-        format!(
-            "{}: SSH authentication failed (check credentials/keys)",
-            context
-        )
-    } else if stderr_lower.contains("could not resolve")
-        || stderr_lower.contains("name or service not known")
-    {
-        format!("{}: unknown host (check hostname)", context)
-    } else if stderr_lower.contains("no such file") || stderr_lower.contains("not a regular file") {
-        format!("{}: remote file not found", context)
-    } else if stderr_lower.contains("timed out") || stderr_lower.contains("connection timed out") {
-        format!("{}: SSH connection timed out", context)
-    } else {
-        format!("{}: {}", context, stderr.trim())
-    }
-}
-
-pub async fn validate_ssh_connection(remote: &RemotePath) -> Result<(), BcmrError> {
-    let target = remote.ssh_target();
-    let output = ssh_command(&target).arg("echo ok").output().await?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(BcmrError::InvalidInput(ssh_error_message(
-            &stderr,
-            &format!("Cannot connect to '{}'", target),
-        )));
-    }
-    Ok(())
-}
-
-pub async fn remote_file_size(remote: &RemotePath) -> Result<Option<u64>, BcmrError> {
-    let output = ssh_command(&remote.ssh_target())
-        .arg(format!(
-            "stat -c '%s' '{}' 2>/dev/null || stat -f '%z' '{}'",
-            shell_escape(&remote.path),
-            shell_escape(&remote.path)
-        ))
-        .output()
-        .await?;
-
-    if !output.status.success() {
-        return Ok(None);
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    Ok(stdout.trim().parse::<u64>().ok())
-}
-
-pub async fn remote_stat(remote: &RemotePath) -> Result<RemoteFileInfo, BcmrError> {
-    let output = ssh_command(&remote.ssh_target())
-        .arg(format!(
-            "stat -c '%F %s' '{}' 2>/dev/null || stat -f '%HT %z' '{}'",
-            shell_escape(&remote.path),
-            shell_escape(&remote.path)
-        ))
-        .output()
-        .await?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(BcmrError::InvalidInput(ssh_error_message(
-            &stderr,
-            &format!("Cannot stat remote path '{}'", remote),
-        )));
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let line = stdout.trim();
-
-    let is_dir = line.to_lowercase().starts_with("directory");
-    let size: u64 = line
-        .rsplit_once(' ')
-        .and_then(|(_, s)| s.parse().ok())
-        .unwrap_or(0);
-
-    Ok(RemoteFileInfo { is_dir, size })
-}
-
-pub async fn remote_total_size(remote: &RemotePath, recursive: bool) -> Result<u64, BcmrError> {
-    let info = remote_stat(remote).await?;
-
-    if !info.is_dir {
-        return Ok(info.size);
-    }
-
-    if !recursive {
-        return Err(BcmrError::InvalidInput(format!(
-            "Remote source '{}' is a directory. Use -r flag for recursive copy.",
-            remote
-        )));
-    }
-
-    let output = ssh_command(&remote.ssh_target())
-        .arg(format!(
-            "find '{}' -type f -exec stat -c '%s' {{}} + 2>/dev/null || find '{}' -type f -exec stat -f '%z' {{}} +",
-            shell_escape(&remote.path), shell_escape(&remote.path)
-        ))
-        .output()
-        .await?;
-
-    if !output.status.success() {
-        return Ok(0);
-    }
-
-    let total: u64 = String::from_utf8_lossy(&output.stdout)
-        .lines()
-        .filter_map(|l| l.trim().parse::<u64>().ok())
-        .sum();
-
-    Ok(total)
-}
-
-pub async fn remote_list_files(remote: &RemotePath) -> Result<Vec<(String, u64, bool)>, BcmrError> {
-    let output = ssh_command(&remote.ssh_target())
-        .arg(format!(
-            "find '{}' -printf '%P\\0%s\\0%y\\0' 2>/dev/null || find '{}' ! -path '{}' -exec stat -f '%N\\0%z\\0%HT\\0' {{}} +",
-            shell_escape(&remote.path), shell_escape(&remote.path), shell_escape(&remote.path)
-        ))
-        .output()
-        .await?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(BcmrError::InvalidInput(ssh_error_message(
-            &stderr,
-            &format!("Cannot list remote directory '{}'", remote),
-        )));
-    }
-
-    let raw = String::from_utf8_lossy(&output.stdout);
-    let fields: Vec<&str> = raw.split('\0').collect();
-    let mut entries = Vec::new();
-
-    let mut i = 0;
-    while i + 2 < fields.len() {
-        let rel_path = fields[i].to_string();
-        let size: u64 = fields[i + 1].parse().unwrap_or(0);
-        let is_dir = fields[i + 2] == "d" || fields[i + 2].to_lowercase().contains("directory");
-        i += 3;
-
-        if rel_path.is_empty() {
-            continue;
-        }
-        entries.push((rel_path, size, is_dir));
-    }
-
-    Ok(entries)
-}
-
 pub struct TransferCallbacks<'a> {
     pub on_progress: &'a (dyn Fn(u64) + Send + Sync),
     pub on_skip: &'a (dyn Fn(u64) + Send + Sync),
     pub on_new_file: &'a (dyn Fn(&str, u64) + Send + Sync),
-}
-
-pub async fn download_file(
-    remote: &RemotePath,
-    local_dst: &Path,
-    cb: TransferCallbacks<'_>,
-    file_size: u64,
-    opts: &RemoteTransferOptions,
-    worker_id: Option<usize>,
-) -> Result<(), BcmrError> {
-    let file_name = remote.path.rsplit('/').next().unwrap_or(&remote.path);
-    (cb.on_new_file)(file_name, file_size);
-
-    if let Some(parent) = local_dst.parent() {
-        if !parent.exists() {
-            tokio::fs::create_dir_all(parent).await?;
-        }
-    }
-
-    let existing_size = if local_dst.exists() {
-        Some(local_dst.metadata()?.len())
-    } else {
-        None
-    };
-    let local_path_for_hash = local_dst.to_path_buf();
-    let remote_for_hash = remote.clone();
-    let remote_for_partial = remote.clone();
-
-    let decision = check_resume_state(
-        opts,
-        existing_size,
-        file_size,
-        async move || {
-            tokio::task::spawn_blocking(move || {
-                crate::core::checksum::calculate_hash(&local_path_for_hash)
-            })
-            .await
-            .map_err(|e| BcmrError::InvalidInput(e.to_string()))?
-            .map_err(BcmrError::Io)
-        },
-        async move || remote_file_hash(&remote_for_hash, None).await,
-        async move |limit| remote_file_hash(&remote_for_partial, Some(limit)).await,
-    )
-    .await?;
-
-    if decision.skip_entirely {
-        (cb.on_skip)(file_size);
-        return Ok(());
-    }
-
-    if decision.skip_bytes > 0 {
-        (cb.on_skip)(decision.skip_bytes);
-    }
-
-    let ssh_cmd = if decision.use_append_mode {
-        format!(
-            "tail -c +{} '{}'",
-            decision.skip_bytes + 1,
-            shell_escape(&remote.path)
-        )
-    } else {
-        format!("cat '{}'", shell_escape(&remote.path))
-    };
-
-    let mut child = make_ssh_cmd(&remote.ssh_target(), worker_id)
-        .arg(ssh_cmd)
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()?;
-
-    let mut stdout = child
-        .stdout
-        .take()
-        .ok_or_else(|| BcmrError::InvalidInput("Failed to capture SSH stdout".to_string()))?;
-    let mut stderr_pipe = child.stderr.take();
-
-    let mut dst_file = if decision.use_append_mode {
-        tokio::fs::OpenOptions::new()
-            .append(true)
-            .open(local_dst)
-            .await?
-    } else {
-        tokio::fs::File::create(local_dst).await?
-    };
-    let mut buffer = vec![0u8; 4 * 1024 * 1024];
-
-    let io_result: Result<(), BcmrError> = async {
-        loop {
-            let n = stdout.read(&mut buffer).await?;
-            if n == 0 {
-                break;
-            }
-            dst_file.write_all(&buffer[..n]).await?;
-            (cb.on_progress)(n as u64);
-        }
-        Ok(())
-    }
-    .await;
-
-    if let Err(e) = io_result {
-        let _ = child.kill().await;
-        if !decision.use_append_mode {
-            let _ = tokio::fs::remove_file(local_dst).await;
-        }
-        return Err(e);
-    }
-
-    let status = child.wait().await?;
-    if !status.success() {
-        let mut stderr_buf = String::new();
-        if let Some(ref mut pipe) = stderr_pipe {
-            let mut buf = Vec::new();
-            let _ = pipe.read_to_end(&mut buf).await;
-            stderr_buf = String::from_utf8_lossy(&buf).to_string();
-        }
-        if !decision.use_append_mode {
-            let _ = tokio::fs::remove_file(local_dst).await;
-        }
-        return Err(BcmrError::InvalidInput(ssh_error_message(
-            &stderr_buf,
-            &format!("Download failed for '{}'", remote),
-        )));
-    }
-
-    if opts.verify {
-        let local_path = local_dst.to_path_buf();
-        let local_hash =
-            tokio::task::spawn_blocking(move || crate::core::checksum::calculate_hash(&local_path))
-                .await
-                .map_err(|e| BcmrError::InvalidInput(e.to_string()))??;
-        let remote_hash = remote_file_hash(remote, None).await?;
-        if local_hash != remote_hash {
-            return Err(BcmrError::InvalidInput(format!(
-                "Verification failed: {} -> '{}'",
-                remote,
-                local_dst.display()
-            )));
-        }
-    }
-
-    if opts.preserve {
-        apply_remote_attrs_locally(remote, local_dst).await?;
-    }
-
-    Ok(())
-}
-
-pub async fn remote_file_hash(
-    remote: &RemotePath,
-    limit: Option<u64>,
-) -> Result<String, BcmrError> {
-    let cmd = match limit {
-        Some(n) => format!("head -c {} '{}'", n, shell_escape(&remote.path)),
-        None => format!("cat '{}'", shell_escape(&remote.path)),
-    };
-
-    let mut child = ssh_command(&remote.ssh_target())
-        .arg(cmd)
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()?;
-
-    let mut stdout = child
-        .stdout
-        .take()
-        .ok_or_else(|| BcmrError::InvalidInput("Failed to capture SSH stdout".to_string()))?;
-
-    let mut hasher = blake3::Hasher::new();
-    let mut buffer = vec![0u8; 4 * 1024 * 1024];
-    loop {
-        let n = stdout.read(&mut buffer).await?;
-        if n == 0 {
-            break;
-        }
-        hasher.update(&buffer[..n]);
-    }
-
-    let status = child.wait().await?;
-    if !status.success() {
-        return Err(BcmrError::InvalidInput(format!(
-            "Failed to hash remote file '{}'",
-            remote
-        )));
-    }
-
-    Ok(hasher.finalize().to_hex().to_string())
-}
-
-pub async fn upload_file(
-    local_src: &Path,
-    remote: &RemotePath,
-    cb: TransferCallbacks<'_>,
-    opts: &RemoteTransferOptions,
-    worker_id: Option<usize>,
-) -> Result<(), BcmrError> {
-    let file_size = local_src.metadata()?.len();
-    let file_name = local_src
-        .file_name()
-        .unwrap_or_default()
-        .to_string_lossy()
-        .to_string();
-    (cb.on_new_file)(&file_name, file_size);
-
-    if let Some(parent) = remote.path.rsplit_once('/') {
-        if !parent.0.is_empty() {
-            let mkdir_out = make_ssh_cmd(&remote.ssh_target(), worker_id)
-                .arg(format!("mkdir -p '{}'", shell_escape(parent.0)))
-                .output()
-                .await?;
-            if !mkdir_out.status.success() {
-                let stderr = String::from_utf8_lossy(&mkdir_out.stderr);
-                return Err(BcmrError::InvalidInput(ssh_error_message(
-                    &stderr,
-                    &format!("Failed to create remote directory '{}'", parent.0),
-                )));
-            }
-        }
-    }
-
-    let existing_size = remote_file_size(remote).await.ok().flatten();
-    let remote_for_hash = remote.clone();
-    let local_path_for_hash = local_src.to_path_buf();
-    let local_path_for_partial = local_src.to_path_buf();
-
-    let decision = check_resume_state(
-        opts,
-        existing_size,
-        file_size,
-        async move || remote_file_hash(&remote_for_hash, None).await,
-        async move || {
-            tokio::task::spawn_blocking(move || {
-                crate::core::checksum::calculate_hash(&local_path_for_hash)
-            })
-            .await
-            .map_err(|e| BcmrError::InvalidInput(e.to_string()))?
-            .map_err(BcmrError::Io)
-        },
-        async move |limit| {
-            tokio::task::spawn_blocking(move || {
-                crate::core::checksum::calculate_partial_hash(&local_path_for_partial, limit)
-            })
-            .await
-            .map_err(|e| BcmrError::InvalidInput(e.to_string()))?
-            .map_err(BcmrError::Io)
-        },
-    )
-    .await?;
-
-    if decision.skip_entirely {
-        (cb.on_skip)(file_size);
-        return Ok(());
-    }
-
-    if decision.skip_bytes > 0 {
-        (cb.on_skip)(decision.skip_bytes);
-    }
-
-    let cat_op = if decision.use_append_mode { ">>" } else { ">" };
-    let mut child = make_ssh_cmd(&remote.ssh_target(), worker_id)
-        .arg(format!("cat {} '{}'", cat_op, shell_escape(&remote.path)))
-        .stdin(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()?;
-
-    let mut stdin = child
-        .stdin
-        .take()
-        .ok_or_else(|| BcmrError::InvalidInput("Failed to capture SSH stdin".to_string()))?;
-
-    let mut src_file = tokio::fs::File::open(local_src).await?;
-
-    if decision.skip_bytes > 0 {
-        use tokio::io::AsyncSeekExt;
-        src_file
-            .seek(std::io::SeekFrom::Start(decision.skip_bytes))
-            .await?;
-    }
-
-    let mut buffer = vec![0u8; 4 * 1024 * 1024];
-
-    let io_result: Result<(), BcmrError> = async {
-        loop {
-            let n = src_file.read(&mut buffer).await?;
-            if n == 0 {
-                break;
-            }
-            stdin.write_all(&buffer[..n]).await?;
-            (cb.on_progress)(n as u64);
-        }
-        Ok(())
-    }
-    .await;
-
-    drop(stdin);
-
-    if let Err(e) = io_result {
-        let _ = child.kill().await;
-        return Err(e);
-    }
-
-    let output = child.wait_with_output().await?;
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(BcmrError::InvalidInput(ssh_error_message(
-            &stderr,
-            &format!("Upload failed for '{}' -> {}", local_src.display(), remote),
-        )));
-    }
-
-    if opts.verify && !verify_remote_file(local_src, remote).await? {
-        return Err(BcmrError::InvalidInput(format!(
-            "Verification failed: '{}' -> {}",
-            local_src.display(),
-            remote
-        )));
-    }
-
-    if opts.preserve {
-        preserve_remote_attrs(local_src, remote).await?;
-    }
-
-    Ok(())
-}
-
-async fn preserve_remote_attrs(local_src: &Path, remote: &RemotePath) -> Result<(), BcmrError> {
-    let meta = local_src.metadata()?;
-
-    let mode = {
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            meta.permissions().mode() & 0o7777
-        }
-        #[cfg(not(unix))]
-        {
-            0o644u32
-        }
-    };
-
-    let mtime_secs = meta
-        .modified()
-        .ok()
-        .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-        .map(|d| d.as_secs())
-        .unwrap_or(0);
-
-    let atime_secs = {
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::MetadataExt;
-            meta.atime() as u64
-        }
-        #[cfg(not(unix))]
-        {
-            mtime_secs
-        }
-    };
-
-    let mtime_ts = unix_to_touch_ts(mtime_secs as i64);
-    let atime_ts = unix_to_touch_ts(atime_secs as i64);
-    let cmd = format!(
-        "TZ=UTC touch -m -t '{}' '{}'; TZ=UTC touch -a -t '{}' '{}'; chmod {:o} '{}'",
-        mtime_ts,
-        shell_escape(&remote.path),
-        atime_ts,
-        shell_escape(&remote.path),
-        mode,
-        shell_escape(&remote.path)
-    );
-    let attr_out = ssh_command(&remote.ssh_target()).arg(cmd).output().await?;
-    if !attr_out.status.success() {
-        let stderr = String::from_utf8_lossy(&attr_out.stderr);
-        return Err(BcmrError::InvalidInput(ssh_error_message(
-            &stderr,
-            &format!("Failed to set attributes on '{}'", remote),
-        )));
-    }
-    Ok(())
-}
-
-async fn get_remote_attrs(remote: &RemotePath) -> Result<(i64, i64, u32), BcmrError> {
-    let output = ssh_command(&remote.ssh_target())
-        .arg(format!(
-            "stat -c '%X %Y %a' '{}' 2>/dev/null || stat -f '%a %m %Lp' '{}'",
-            shell_escape(&remote.path),
-            shell_escape(&remote.path)
-        ))
-        .output()
-        .await?;
-
-    if !output.status.success() {
-        return Err(BcmrError::InvalidInput(format!(
-            "Cannot stat remote file '{}'",
-            remote
-        )));
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let parts: Vec<&str> = stdout.split_whitespace().collect();
-    let atime_secs: i64 = parts.first().and_then(|s| s.parse().ok()).unwrap_or(0);
-    let mtime_secs: i64 = parts.get(1).and_then(|s| s.parse().ok()).unwrap_or(0);
-    let mode: u32 = parts
-        .get(2)
-        .and_then(|s| u32::from_str_radix(s, 8).ok())
-        .unwrap_or(0o644);
-
-    Ok((atime_secs, mtime_secs, mode))
-}
-
-fn apply_local_attrs(
-    local_path: &Path,
-    atime_secs: i64,
-    mtime_secs: i64,
-    _mode: u32,
-) -> Result<(), BcmrError> {
-    let atime = filetime::FileTime::from_unix_time(atime_secs, 0);
-    let mtime = filetime::FileTime::from_unix_time(mtime_secs, 0);
-    filetime::set_file_times(local_path, atime, mtime)?;
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        std::fs::set_permissions(local_path, std::fs::Permissions::from_mode(_mode))?;
-    }
-
-    Ok(())
-}
-
-async fn apply_remote_attrs_locally(
-    remote: &RemotePath,
-    local_path: &Path,
-) -> Result<(), BcmrError> {
-    let (atime_secs, mtime_secs, mode) = get_remote_attrs(remote).await?;
-    apply_local_attrs(local_path, atime_secs, mtime_secs, mode)?;
-    Ok(())
-}
-
-pub async fn verify_remote_file(local_src: &Path, remote: &RemotePath) -> Result<bool, BcmrError> {
-    use crate::core::checksum;
-
-    let local_path = local_src.to_path_buf();
-    let local_hash = tokio::task::spawn_blocking(move || checksum::calculate_hash(&local_path))
-        .await
-        .map_err(|e| BcmrError::InvalidInput(e.to_string()))??;
-
-    let remote_hash = remote_file_hash(remote, None).await?;
-
-    Ok(local_hash == remote_hash)
-}
-
-fn unix_to_touch_ts(secs: i64) -> String {
-    let days = secs.div_euclid(86400);
-    let rem = secs.rem_euclid(86400);
-    let hours = rem / 3600;
-    let minutes = (rem % 3600) / 60;
-    let seconds = rem % 60;
-
-    let z = days as i32 + 719468;
-    let era = if z >= 0 { z } else { z - 146096 } / 146097;
-    let doe = (z - era * 146097) as u32;
-    let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
-    let y = yoe as i32 + era * 400;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2) / 153;
-    let d = doy - (153 * mp + 2) / 5 + 1;
-    let m = if mp < 10 { mp + 3 } else { mp - 9 };
-    let y = if m <= 2 { y + 1 } else { y };
-
-    format!(
-        "{:04}{:02}{:02}{:02}{:02}.{:02}",
-        y, m, d, hours, minutes, seconds
-    )
-}
-
-pub async fn download_directory(
-    remote: &RemotePath,
-    local_dst: &Path,
-    cb: TransferCallbacks<'_>,
-    excludes: &[regex::Regex],
-    opts: &RemoteTransferOptions,
-) -> Result<(), BcmrError> {
-    let entries = remote_list_files(remote).await?;
-
-    let entries: Vec<_> = entries
-        .into_iter()
-        .filter(|(rel_path, _, _)| {
-            !crate::core::traversal::is_excluded(std::path::Path::new(rel_path), excludes)
-        })
-        .collect();
-
-    for (rel_path, _, is_dir) in &entries {
-        if *is_dir {
-            let dir_path = local_dst.join(rel_path);
-            if !dir_path.exists() {
-                tokio::fs::create_dir_all(&dir_path).await?;
-            }
-        }
-    }
-
-    for (rel_path, size, is_dir) in &entries {
-        if *is_dir {
-            continue;
-        }
-        let file_remote = RemotePath {
-            user: remote.user.clone(),
-            host: remote.host.clone(),
-            path: format!("{}/{}", remote.path, rel_path),
-        };
-        let local_file = local_dst.join(rel_path);
-        download_file(
-            &file_remote,
-            &local_file,
-            TransferCallbacks {
-                on_progress: cb.on_progress,
-                on_skip: cb.on_skip,
-                on_new_file: cb.on_new_file,
-            },
-            *size,
-            opts,
-            None,
-        )
-        .await?;
-    }
-
-    Ok(())
-}
-
-pub async fn ensure_remote_tree(local_src: &Path, remote: &RemotePath) -> Result<(), BcmrError> {
-    use crate::core::traversal;
-
-    ssh_command(&remote.ssh_target())
-        .arg(format!("mkdir -p '{}'", shell_escape(&remote.path)))
-        .output()
-        .await?;
-
-    let excludes: Vec<regex::Regex> = Vec::new();
-    let mut dirs = Vec::new();
-
-    for entry in traversal::walk(local_src, true, false, 1, &excludes) {
-        let entry = entry?;
-        let path = entry.path();
-        if path.is_dir() {
-            let rel = path.strip_prefix(local_src)?;
-            dirs.push(rel.to_path_buf());
-        }
-    }
-
-    if !dirs.is_empty() {
-        let mkdir_cmd = dirs
-            .iter()
-            .map(|d| {
-                format!(
-                    "'{}/{}'",
-                    shell_escape(&remote.path),
-                    shell_escape(&d.display().to_string())
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(" ");
-
-        ssh_command(&remote.ssh_target())
-            .arg(format!("mkdir -p {}", mkdir_cmd))
-            .output()
-            .await?;
-    }
-
-    Ok(())
-}
-
-pub async fn upload_directory(
-    local_src: &Path,
-    remote: &RemotePath,
-    cb: TransferCallbacks<'_>,
-    excludes: &[regex::Regex],
-    opts: &RemoteTransferOptions,
-) -> Result<(), BcmrError> {
-    use crate::core::traversal;
-
-    let output = ssh_command(&remote.ssh_target())
-        .arg(format!("mkdir -p '{}'", shell_escape(&remote.path)))
-        .output()
-        .await?;
-
-    if !output.status.success() {
-        return Err(BcmrError::InvalidInput(format!(
-            "Failed to create remote directory '{}'",
-            remote
-        )));
-    }
-
-    let mut dirs = Vec::new();
-    let mut files = Vec::new();
-
-    for entry in traversal::walk(local_src, true, false, 1, excludes) {
-        let entry = entry?;
-        let path = entry.path();
-        let rel = path.strip_prefix(local_src)?;
-
-        if path.is_dir() {
-            dirs.push(rel.to_path_buf());
-        } else if path.is_file() {
-            files.push((path.to_path_buf(), rel.to_path_buf()));
-        }
-    }
-
-    if !dirs.is_empty() {
-        let mkdir_cmd = dirs
-            .iter()
-            .map(|d| {
-                format!(
-                    "'{}/{}'",
-                    shell_escape(&remote.path),
-                    shell_escape(&d.display().to_string())
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(" ");
-
-        ssh_command(&remote.ssh_target())
-            .arg(format!("mkdir -p {}", mkdir_cmd))
-            .output()
-            .await?;
-    }
-
-    for (local_path, rel_path) in &files {
-        let file_remote = RemotePath {
-            user: remote.user.clone(),
-            host: remote.host.clone(),
-            path: format!("{}/{}", remote.path, rel_path.display()),
-        };
-        upload_file(
-            local_path,
-            &file_remote,
-            TransferCallbacks {
-                on_progress: cb.on_progress,
-                on_skip: cb.on_skip,
-                on_new_file: cb.on_new_file,
-            },
-            opts,
-            None,
-        )
-        .await?;
-    }
-
-    Ok(())
-}
-
-pub async fn complete_remote_path(partial: &str) -> Vec<String> {
-    let remote = match parse_remote_path(partial) {
-        Some(r) => r,
-        None => return Vec::new(),
-    };
-
-    let (dir, prefix) = if remote.path.ends_with('/') || remote.path == "." {
-        (remote.path.clone(), String::new())
-    } else if let Some(pos) = remote.path.rfind('/') {
-        (
-            remote.path[..=pos].to_string(),
-            remote.path[pos + 1..].to_string(),
-        )
-    } else {
-        (".".to_string(), remote.path.clone())
-    };
-
-    let target = remote.ssh_target();
-    let output = match ssh_command(&target)
-        .arg(format!("ls -1ap '{}' 2>/dev/null", shell_escape(&dir)))
-        .output()
-        .await
-    {
-        Ok(o) if o.status.success() => o,
-        _ => return Vec::new(),
-    };
-
-    let base = if dir == "." { String::new() } else { dir };
-
-    String::from_utf8_lossy(&output.stdout)
-        .lines()
-        .filter(|l| *l != "./" && *l != "../")
-        .filter(|l| prefix.is_empty() || l.starts_with(&prefix))
-        .map(|l| format!("{}:{}{}", target, base, l))
-        .collect()
 }
 
 #[cfg(test)]
@@ -1220,97 +211,11 @@ mod tests {
 
     #[test]
     fn test_shell_escape_no_quotes() {
-        assert_eq!(shell_escape("simple"), "simple");
+        assert_eq!(super::ssh_cmd::shell_escape("simple"), "simple");
     }
 
     #[test]
     fn test_shell_escape_with_quotes() {
-        assert_eq!(shell_escape("it's"), "it'\\''s");
-    }
-
-    #[test]
-    fn test_unix_to_touch_ts_epoch() {
-        assert_eq!(unix_to_touch_ts(0), "197001010000.00");
-    }
-
-    #[test]
-    fn test_unix_to_touch_ts_known_date() {
-        let ts = unix_to_touch_ts(1577880000);
-        assert_eq!(ts, "202001011200.00");
-    }
-
-    #[test]
-    fn test_unix_to_touch_ts_with_seconds() {
-        let ts = unix_to_touch_ts(1592210445);
-        assert!(ts.ends_with(".45"));
-    }
-
-    #[tokio::test]
-    async fn test_check_resume_state_same_size_requires_hash_match() {
-        let opts = RemoteTransferOptions {
-            resume: true,
-            ..Default::default()
-        };
-
-        let decision = check_resume_state(
-            &opts,
-            Some(1024),
-            1024,
-            async || Ok("existing".to_string()),
-            async || Ok("source".to_string()),
-            async |_| Ok("unused".to_string()),
-        )
-        .await
-        .expect("decision should compute");
-
-        assert!(!decision.skip_entirely);
-        assert_eq!(decision.skip_bytes, 0);
-        assert!(!decision.use_append_mode);
-    }
-
-    #[tokio::test]
-    async fn test_check_resume_state_shorter_prefix_requires_hash_match() {
-        let opts = RemoteTransferOptions {
-            append: true,
-            ..Default::default()
-        };
-
-        let decision = check_resume_state(
-            &opts,
-            Some(512),
-            1024,
-            async || Ok("corrupt-prefix".to_string()),
-            async || Ok("unused".to_string()),
-            async |_| Ok("source-prefix".to_string()),
-        )
-        .await
-        .expect("decision should compute");
-
-        assert!(!decision.skip_entirely);
-        assert_eq!(decision.skip_bytes, 0);
-        assert!(!decision.use_append_mode);
-    }
-
-    #[tokio::test]
-    async fn test_check_resume_state_matching_prefix_allows_append() {
-        let opts = RemoteTransferOptions {
-            resume: true,
-            ..Default::default()
-        };
-
-        let decision = check_resume_state(
-            &opts,
-            Some(512),
-            1024,
-            async || Ok("prefix-hash".to_string()),
-            async || Ok("full-hash".to_string()),
-            async |_| Ok("prefix-hash".to_string()),
-        )
-        .await
-        .expect("decision should compute");
-
-        assert!(!decision.skip_entirely);
-        assert_eq!(decision.skip_bytes, 512);
-        assert!(decision.use_append_mode);
+        assert_eq!(super::ssh_cmd::shell_escape("it's"), "it'\\''s");
     }
 }
